Remove structs from request to model
